services/sr_analysis/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "analysis-service"
version = "0.1.0"
description = "SEO analysis microservice — scores target pages against SERP competitors"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Optimization Agent Team"},
]
dependencies = [
    "fastapi>=0.135.0",
    "google-genai>=1.73.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.13.0",
    "PyYAML>=6.0",
    "structlog>=25.0.0",
    "uvicorn>=0.44.0",
    "httpx>=0.28.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
plugins = ["pydantic.mypy"]
strict = true
